+++ packages/nuxt/src/module.ts
@@ -1,6 +1,7 @@
-import { addPlugin, createResolver, defineNuxtModule } from '@nuxt/kit';
+import * as fs from 'fs';
+import * as path from 'path';
+import { addPlugin, addPluginTemplate, createResolver, defineNuxtModule } from '@nuxt/kit';
 import type { SentryNuxtOptions } from './common/types';
-import { findDefaultSdkInitFile, injectSentryConfigImport } from './common/utils';
 
 export type ModuleOptions = SentryNuxtOptions;
 
@@ -20,19 +21,40 @@ export default defineNuxtModule<ModuleOptions>({
     const clientConfigFile = findDefaultSdkInitFile('client');
 
     if (clientConfigFile) {
-      nuxt.options.alias['#sentry-client-config'] = buildDirResolver.resolve(`/${clientConfigFile}`);
+      // Inject the client-side Sentry config file with a side effect import
+      addPluginTemplate({
+        mode: 'client',
+        filename: 'sentry-client-config.mjs',
+        getContents: () =>
+          `import "${buildDirResolver.resolve(`/${clientConfigFile}`)}"\n` +
+          'export default defineNuxtPlugin(() => {})',
+      });
 
-      injectSentryConfigImport('#sentry-client-config');
-
-      addPlugin(moduleDirResolver.resolve('./runtime/plugins/sentry.client'));
+      addPlugin({ src: moduleDirResolver.resolve('./runtime/plugins/sentry.client'), mode: 'client' });
     }
 
     const serverConfigFile = findDefaultSdkInitFile('server');
 
     if (serverConfigFile) {
-      nuxt.options.alias['#sentry-server-config'] = buildDirResolver.resolve(`/${serverConfigFile}`);
-
-      injectSentryConfigImport('#sentry-server-config');
+      // Inject the server-side Sentry config file with a side effect import
+      addPluginTemplate({
+        mode: 'server',
+        filename: 'sentry-server-config.mjs',
+        getContents: () =>
+          `import "${buildDirResolver.resolve(`/${serverConfigFile}`)}"\n` +
+          'export default defineNuxtPlugin(() => {})',
+      });
     }
   },
 });
+
+function findDefaultSdkInitFile(type: 'server' | 'client'): string | undefined {
+  const possibleFileExtensions = ['ts', 'js', 'mjs', 'cjs', 'mts', 'cts'];
+
+  const cwd = process.cwd();
+  const filePath = possibleFileExtensions
+    .map(e => path.resolve(path.join(cwd, `sentry.${type}.config.${e}`)))
+    .find(filename => fs.existsSync(filename));
+
+  return filePath ? path.basename(filePath) : undefined;
+}
